ndisc: don't call solicit_routers() from clean_dns_*() functions
This was done since NDisc code was added to NetworkManager in
commit c3a4656a68f9 ('rdisc: libndp implementation').

Note what it does: in clean_dns_*() we will call solicit_router()
if the half-life of any entity is expired. That doesn't seem right.
Why only for dns_servers and dns_domains, but not routes, addresses
and gateways?

Also, why would the timings for when we solicit depend on when
elements expire. It is "normal" that some of them will expire.
We should solicit based on other parameters, like keeping track
of when and how to solicit.

Note that there is a change in behavior here: if we stopped
soliciting (either because we received our first RA or because
we run out of retries), then we now will never start again.
Previously this was a mechanism so that we would eventually
start solicitating again. But that is probably fine. RFC4861
([1]) says

   If a host sends MAX_RTR_SOLICITATIONS solicitations, and receives no
   Router Advertisements after having waited MAX_RTR_SOLICITATION_DELAY
   seconds after sending the last solicitation, the host concludes that
   there are no routers on the link for the purpose of [ADDRCONF].
   However, the host continues to receive and process Router
   Advertisements messages in the event that routers appear on the link.

Likewise, it seems systemd has no mechanism for restarting solicitation
after receiving the first RA ([2]).

This will be addressed soon in a follow-up commit.

[1] https://tools.ietf.org/html/rfc4861#section-6.3.7
[2] https://github.com/systemd/systemd/blob/9b25429cb4e9b71e25cd9094ff2bbe3351aa9329/src/libsystemd-network/sd-ndisc.c#L251